python/requirements.txt
customtkinter
Pillow
qrcode
pyserial
brother_ql
